gs api

+++ src/app/prescreening/prescreening.component.ts
@@ -22,6 +22,7 @@ export class PrescreeningComponent {
   ) { }
 
   ngOnInit() {
+    this.fetchLocalIp();
   }
 
   categorySelected(categoryId: string) {
@@ -90,57 +91,103 @@ export class PrescreeningComponent {
     }
   }
 
-  onSubmit() {
+  async onSubmit() {
 
+    // Check if the IP Address exists, 
+    // if it does, we redirect to not-eligible.
+    const ipExists = await this.checkIfIpExists(this.ipAddress);
 
+    if (ipExists) {
+      return;
+    }
+
+    // If the IP Address does not exist, we log it, before we do anything else, 
+    // Do do it to so that user cannot submit anything else again.
+    await this.logIp(this.ipAddress);
+
+    // If the user selects 'None of the above', we redirect to not-eligible,
+    // and also remember on client side that the user is not eligible. So that
+    // even if the user uses VPN to change IP Address, we still remember that.
     if (this.categoryId === 'none-of-the-above') {
-
       localStorage.setItem('ps', JSON.stringify(true));
-      // fetch the IP address
-      this.logIp(this.fetchLocalIp());
       this.router.navigate(['/not-eligible']);
 
     } else {
-      
+
+      // If the user is eligible, we first make sure that the we had not
+      // previously marked the user as not eligible.
       if (this.checkNotEligible()) {
         this.router.navigate(['/category-full']);
+
       } else {
         this.router.navigate(['/consent-and-demographic']);
-      }
-      
-    }
 
-  } 
+      }
+    }
+  }
+  
 
   checkNotEligible(): boolean {
     const value = localStorage.getItem('ps');
 
     if (value === 'true') {
       return true;
-
     } else {
       return false;
-
     }
   }
 
+
+  /* 
+  Fetches the client's IP Address.
+  */
   fetchLocalIp():any {
     // Fetch the IP address
     this.http.get<{origin: string}>('https://httpbin.org/ip')
       .subscribe(response => {
         this.ipAddress = response.origin;
-        return this.ipAddress;
       }, error => {
         console.error('Error fetching IP address:', error);
       });
   }
 
-  logIp(ipAddress: string) {
+
+  /*
+  Logs the IP Address in the database.
+  */
+  logIp(ipAddress: string): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
     this.sheetsService.logIpAddress(ipAddress).subscribe(response => {
-      console.log('IP Address logged', response);
+      // console.log('IP Address logged', response);
+      resolve();
     }, error => {
-      console.error('Error logging IP Address', error);
+      // console.error('Error logging IP Address', error);
+      reject();
+    });
     });
   }
+
+
+  /*
+  Checks if the IP Address exists in the database, if it does, redirect to category full.
+  */
+  checkIfIpExists(ipAddress: string): Promise<boolean> {
+    return new Promise((resolve, reject) => {
+    this.sheetsService.checkIfIpExists(ipAddress).subscribe((response: any) => {
+        if (response.exists) {
+          // console.log('IP Address exists');
+          this.router.navigate(['/category-full']);
+          resolve(true);
+        } else {
+          resolve(false);
+        }
+
+    }, error => {
+        // console.error('Error checking if IP Address exists', error);
+        reject(error);
+    });
+    });
+  }
+
 
 }
